move `Payload` into `fixtures.rs`
