+++ lib/blob_store/benches/bustle_bench/fixture.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use std::sync::{Arc, OnceLock};
 
-use blob_store::payload::Payload;
+use blob_store::fixtures::Payload;
 use bustle::CollectionHandle;
 use parking_lot::RwLock;
 use serde_json::json;
