+++ lib/blob_store/src/blob_store.rs
@@ -449,8 +449,7 @@ mod tests {
     use crate::config::{
         DEFAULT_BLOCK_SIZE_BYTES, DEFAULT_PAGE_SIZE_BYTES, DEFAULT_REGION_SIZE_BLOCKS,
     };
-    use crate::fixtures::{empty_storage, empty_storage_sized, random_payload, HM_FIELDS};
-    use crate::payload::Payload;
+    use crate::fixtures::{empty_storage, empty_storage_sized, random_payload, Payload, HM_FIELDS};
 
     #[test]
     fn test_empty_payload_storage() {
